feat: add alternative processing driver "redis-queue" (rq)
allows different env var usage for different containers, e.g. separate ollama URLs for different Gitlab domains
also prevents LLM calls from timing out due to sequential queue processing
