+++ .env.dist
@@ -21,7 +21,8 @@ DEEPSEEK_API_BASE_URL=https://api.deepseek.com
 DEEPSEEK_API_MODEL=deepseek-chat
 
 #OllaMA settings; 注意: 如果使用 Docker 部署，127.0.0.1 指向的是容器内部的地址。请将其替换为实际的 Ollama服务器IP地址。
-OLLAMA_API_BASE_URL=http://127.0.0.1:11434
+#OLLAMA_API_BASE_URL=http://127.0.0.1:11434
+OLLAMA_API_BASE_URL=http://host.docker.internal:11434
 OLLAMA_API_MODEL=deepseek-r1:latest
 
 #支持review的文件类型
@@ -59,3 +60,9 @@ PUSH_REVIEW_ENABLED=1
 # Dashboard登录用户名和密码
 DASHBOARD_USER=admin
 DASHBOARD_PASSWORD=admin
+
+# queue (async, rq)
+QUEUE_DRIVER=async
+REDIS_HOST=redis
+# REDIS_HOST=127.0.0.1
+# REDIS_PORT=6379
