+++ docker-compose.prod.yml
@@ -13,8 +13,35 @@ services:
       - log:/app/log
     env_file:
       - .env
+    depends_on:
+      redis:
+        condition: service_started
+    restart: unless-stopped
+
+  worker:
+    build:
+      context: .
+      dockerfile: Dockerfile
+      target: worker
+    # image: sunmh207/ai-codereview-gitlab:1.2.3-worker
+    volumes_from:
+      - app
+    env_file:
+      - .env
+    depends_on:
+      redis:
+        condition: service_started
+    restart: unless-stopped
+
+  redis:
+    image: redis:alpine
+#    ports:
+#      - "${REDIS_PORT}:6379"
+    volumes:
+      - redis_data:/data
     restart: unless-stopped
 
 volumes:
   data:
   log:
+  redis_data:
