+++ requirements.txt
@@ -8,3 +8,4 @@ httpx[socks]
 ollama==0.4.7
 pandas==2.2.3
 streamlit==1.42.2
+rq==2.1.0
