+++ supervisord.worker-default.conf
@@ -1,0 +1,15 @@
+[supervisord]
+nodaemon=true
+user=root
+
+[program:worker]
+command=rq worker default --url redis://redis:6379 --path /app
+autostart=true
+autorestart=true
+numprocs=1
+stdout_logfile=/dev/stdout
+stderr_logfile=/dev/stderr
+stdout_maxbytes=0
+stderr_maxbytes=0
+stdout_logfile_maxbytes = 0
+stderr_logfile_maxbytes = 0
